refactor: add useJobs
